requirements.txt
python-twitter
python-dotenv==0.14.0
oauth2
gunicorn
discord.py==1.3.4
quart
pony